requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-crispy-forms==1.13.0
django-filter==21.1
django-login-required-middleware==0.7.0
django-widget-tweaks==1.4.9
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
